+++ lib/rbcli/components/config/backends/yaml.rb
@@ -12,9 +12,7 @@ class Rbcli::UserConf::Yaml < Rbcli::UserConf::Backend
     begin
       parsed_str = YAML.safe_load(str, symbolize_names: true, aliases: true, permitted_classes: [Symbol])
     rescue Psych::SyntaxError, Psych::DisallowedClass => e
-      Rbcli.log.warn "Invalid #{@type} syntax found at '#{@path}'", "CONF"
-      Rbcli.log.warn e.message, "CONF"
-      Hash.new
+      raise Rbcli::ParseError.new "Invalid #{@type} syntax found at '#{@path}': #{e.message}"
     else
       @loaded = true
       parsed_str